+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Initial support for BOLT12 offers
 
+- Improve the main screen initial loading times
+
 ## 0.2.17 - 2020-09-03
 
 Hotfix release - Fix YAML rendering of payments with unknown destination/description. This bug only manfiests when expert mode is enabled.